refactor: 엔티티 레이어의 enum 타입 위치를 도메인 영역으로 변경 (#36)
