+++ src/main/java/com/flytrap/rssreader/domain/member/Member.java
@@ -1,6 +1,5 @@
 package com.flytrap.rssreader.domain.member;
 
-import com.flytrap.rssreader.infrastructure.entity.member.OauthServer;
 import java.io.Serializable;
 import java.time.Instant;
 import lombok.AccessLevel;
