+++ src/main/java/com/flytrap/rssreader/domain/member/OauthServer.java
@@ -1,0 +1,5 @@
+package com.flytrap.rssreader.domain.member;
+
+public enum OauthServer {
+    GITHUB;
+}
