+++ src/main/java/com/flytrap/rssreader/domain/subscribe/BlogPlatform.java
@@ -1,4 +1,4 @@
-package com.flytrap.rssreader.infrastructure.entity.subscribe;
+package com.flytrap.rssreader.domain.subscribe;
 
 public enum BlogPlatform {
     VELOG, TISTORY, GITPAGE, ETC
