+++ src/main/java/com/flytrap/rssreader/infrastructure/entity/member/MemberEntity.java
@@ -1,6 +1,7 @@
 package com.flytrap.rssreader.infrastructure.entity.member;
 
 import com.flytrap.rssreader.domain.member.Member;
+import com.flytrap.rssreader.domain.member.OauthServer;
 import com.flytrap.rssreader.infrastructure.entity.subscribe.SubscribeEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
