--- src/main/java/com/flytrap/rssreader/infrastructure/entity/member/OauthServer.java
@@ -1,5 +1,0 @@
-package com.flytrap.rssreader.infrastructure.entity.member;
-
-public enum OauthServer {
-    GITHUB;
-}
